Very basic start on supporting gophers (gopher over tls)
Currently this works with a new enough curl but doesn't handle connection
fallback (the standard seems to now be falling back if gophers doesn't work)
or handling gophers in gopher listings. I think we need connection fallback
first as that way, if gophers is used to connect to a host, we can try
gophers for the selectors and fall back if it doesn't work. I have no idea
how to handle submitting input across hosts when we end up falling back to
gopher but I don't think anyone's asked about that in the community yet.