Make secure to insecure form submission behaviour the same for https and gophers
Close gaps where someone either sets up a form to submit from a gophers
connection to a gopher or http host or, probably only theoretically, an
https host to a gopher one. However theoretical this is, we should just not
have this gap and this change hopefully fixes that.

Note that, until we implement proper gophers handling in gopher listing
transformation this'll temporarily break form submissions for gophers hosts.