First small step toward mail forwarding, which has never been implemented or even approached before.
fwd: is suppose to be like attach: at a high level.
Check the mime standard and change the words so it looks like a forwarding not an attachment. I hope it's not a big deal.
The file or session referenced should be a raw email, and we should probably check for that.
Eventually an fwd edbrowse command will pull the raw email out of the background, put it in an empty session, and include fwd: x
Switch from browse to text, you type your stuff, then sm.
So it's like re, but not like re.
This will be "in my spare time" unless someone else wants to run with it.