Add copyString function to replace strncpy usage which generates warnings
The function uses the POSIX memccpy (yes they have two different functions
where the spelling of the function names is a doubled letter in the middle)
so we don't extra pad buffers. The copyString function also null-terminates the
copied strings.

Also alter other code in stringfile to use the new function and correct the
readlink usage according to the POSIX standard.