Another attempt to fix getting elements by id
The last commit wasn't right re: what the this object was or where the id
hash lived. Correct that as well as fixing the actual issue which is that
the gebi function had no this defined and needed to use the correct document
object as it's not part of the prototype.