Wrap vasprintf() and asprintf() to check return codes similar to allocMem()
It's probably very rare that they will fail but it's a good idea to check.
This also standardises various checks on the return codes which used various
error messages, some of which may have been misleading. We also now add the
error (using strerror) to the message in case something odd happens.

The two functions are createFormattedStringV() for vasprintf() and createFormattedString() for asprintf().