Remove now unnecessary and potentially confusing debug message around downloads
We already get the content type at db3 and the size is now part of the
download prompt if displayed. As such the extra debug message (also only
displayed with the download prompt) is now unnecessary. It's also become a
little confusing since the decision is made in the header callback and so
only one of content type and content length may have been parsed when the
decision to prompt is made. Extra logic could have been added to print a
better message but this would add complexity for the sake of a debug message
where all the data is already available.