Improvements to structuredClone and postMessage
Postmessage needs to use structuredClone on the message as it can be
anything. There's also a case where we don't give a target origin which we
now handle, I think correctly.

structuredClone now gets bound to the window so we can run properly in
frames. The type checking has also been improved to account for structured
cloning across frames (needed by postMessage).