Rewrite custom elements registry to use a map rather than a plain object
Since it appears that we're only supposed to validate names on definition
(according to code in the wild, specifically as part of
https://kb.synology.com/ which does a get on an invalid name and fails to
load if that blows up) use a map to hold the keys and values rather than a
plain object. This means there's no risk of names colliding with object
properties so the invalid names just get undefined which appears to be
expected.